feat(onboarding): first-access setup gate + LLM provider wizard
